got window working

+++ src/graphics/output.rs
@@ -6,8 +6,8 @@ use crate::graphics::color::Color;
 pub struct RenderOutputCoords(pub i32, pub i32);
 
 pub trait RenderOutputter {
-    fn width(&self) -> u16;
-    fn height(&self) -> u16;
+    fn width(&self) -> usize;
+    fn height(&self) -> usize;
     fn set(&mut self, x: i32, y: i32, color: &Color);
     fn render(&mut self) -> Result<(), Box<dyn Error>>;
 }
